+++ src/app.ts
@@ -12,16 +12,16 @@ import { i18next, i18nextMiddleware } from './middleware/translation';
 import { rateLimiter } from './middleware/rate-limiter';
 import session from './middleware/session';
 import { requestContext } from './middleware/context';
-import { authRouter } from './route/auth';
-import { healthcheckRouter } from './route/healthcheck';
-import { datasetRouter } from './route/dataset';
-import { errorHandler } from './route/error-handler';
-import { providerRouter } from './route/provider';
-import { topicRouter } from './route/topic';
-import { organisationRouter } from './route/organisation';
-import { teamRouter } from './route/team';
-import { translationRouter } from './route/translation';
-import { consumerRouter } from './route/consumer';
+import { authRouter } from './routes/auth';
+import { healthcheckRouter } from './routes/healthcheck';
+import { datasetRouter } from './routes/dataset';
+import { errorHandler } from './routes/error-handler';
+import { providerRouter } from './routes/provider';
+import { topicRouter } from './routes/topic';
+import { organisationRouter } from './routes/organisation';
+import { teamRouter } from './routes/team';
+import { translationRouter } from './routes/translation';
+import { consumerRouter } from './routes/consumer';
 import { initServices } from './middleware/services';
 
 const app: Application = express();
